Update ManagedSubscribe.java to handle termination
When terminating, if the connection is closed, immediate resub results in the following error 

'An active managed subscription with the same ID already exists. If a previous subscription with the same ID was stopped, there can be a delay before it expires. Try subscribing again later.'.   

This change aims to allow ctrl+c gracefully closes the connection.